// src/enums/Privilege.ts
enum Privilege {

  ASK_BUDDY_EDIT = "ASK_BUDDY_EDIT",
  ASK_BUDDY_VIEW = "ASK_BUDDY_VIEW",
  ASK_BUDDY_ANALYTICS = "ASK_BUDDY_ANALYTICS",

  COMM_AI_EDIT = "COMM_AI_EDIT",
  COMM_AI_VIEW = "COMM_AI_VIEW",
  COMM_AI_ANALYTICS = "COMM_AI_ANALYTICS",

  RESP_AI_EDIT = "RESP_AI_EDIT",
  RESP_AI_VIEW = "RESP_AI_VIEW",
  RESP_AI_ANALYTICS = "RESP_AI_ANALYTICS",

  MARKET_AI_EDIT = "MARKET_AI_EDIT",
  MARKET_AI_VIEW = "MARKET_AI_VIEW",
  MARKET_AI_ANALYTICS = "MARKET_AI_ANALYTICS",

  PROF_EDGE_EDIT = "PROF_EDGE_EDIT",
  PROF_EDGE_VIEW = "PROF_EDGE_VIEW",
  PROF_EDGE_ANALYTICS = "PROF_EDGE_ANALYTICS",

  CLIENT_SETUP_EDIT = "CLIENT_SETUP_EDIT",
  CLIENT_SETUP_VIEW = "CLIENT_SETUP_VIEW",
  CLIENT_SETUP_ANALYTICS = "CLIENT_SETUP_ANALYTICS",

  USER_ADMIN_EDIT = "USER_ADMIN_EDIT",
  USER_ADMIN_VIEW = "USER_ADMIN_VIEW",
  USER_ADMIN_ANALYTICS = "USER_ADMIN_ANALYTICS",

  BILLING_EDIT = "BILLING_EDIT",
  BILLING_VIEW = "BILLING_VIEW",
  BILLING_ANALYTICS = "BILLING_ANALYTICS",

  NPS_ANALYTICS = "NPS_ANALYTICS",
  NPS_EDIT = "NPS_EDIT",
  NPS_VIEW = "NPS_VIEW",

  INSIGHT_ANALYTICS = "INSIGHT_ANALYTICS",

  //only if user.isAdmin = true; cannot be part of any role at this point
  SYSTEM_ADMIN_EDIT = "SYSTEM_ADMIN_EDIT"
}

export default Privilege